+++ src/player_car.h
@@ -14,7 +14,7 @@ extern uint16_t player_car_current_track_segment_changes_applied;
 extern int32_t camera_track_position;
 extern int32_t player_car_track_position;
 extern int32_t player_car_logical_xpos;
-extern int32_t player_car_speed;
+extern int16_t player_car_speed;
 extern int32_t player_car_steering;
 extern uint16_t player_car_state;
 extern uint16_t player_car_invincible_countdown;
